HYD-5231 Fix catch all route in IML router
The router is throwing an error in a special case instead of routing to
the wildcard route. The issue occurs when you define a route using a specific method and also define a wildcard route for all methods. When making a request using the same path but with a different method, the wildcard route doesn't pick up the request and an error is thrown instead. For example, take the following 2 routes:

router.post('/api/mock')
router.all('(.*)')

Notice that the /api/mock path is defined for a POST. The problem will
occur if I call:

router.go('/api/mock', {verb: 'get'})

Instead of going to the wildcard route, it throws an error. This ticket
addresses the issue so that the wildcard will handle the route
appropriately.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>