HYD-5304 Update router to invoke an optional callback if it is passed in at the end of the pipeline
While working on HYD-5096 (Stub Daddy Update) we discovered that it
would be beneficial for the router to invoke an optional callback with
the request, response, and data at the end of the pipeline. This allows
the inline service to get the response after the request completes all
middleware.

This patch will add the optional callback to the list of pipes (if it's
provided of course), where it will be invoked with the parameters that
are passed in by the last middleware.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>